--- STEP3/main.swift
@@ -1,8 +1,0 @@
-//
-//  main.swift
-//  CodeStarterCamp_Week1
-//
-//  Created by 한겨레 on 2022/06/23.
-//
-
-import Foundation
